significant changes throughout including gaming wallet staking, update toast messages, multiplier fix - i think i've fixed a memory leak - i think i fixed the issue with the win animation playing sometimes when a user spins (related to changing tabs during a win) -  added a "stake" tab to the main page - new header bar to make it easier to navigate between game and house - leaderboard cleanup/improvements that we discussed today - added ability to manage your Gaming Wallet stake from the house page, in addition to an externally connected wallet - added "new version" toast message which will prompt a user to refresh when a new version is released - removed stats from game queue (confusing and redundant) - removed Analytics tab from house (confusing -- need to rethink) - hardcoded multipliers to fix issue loading multipliers from contract - fixed issue with leaderboard sometimes re-sorting out of order

+++ hov.sql
@@ -1088,8 +1088,8 @@ CREATE OR REPLACE FUNCTION get_payline_analysis(
 RETURNS TABLE(
     paylines_count INTEGER,
     total_bets BIGINT,
-    total_amount_bet BIGINT,
-    total_amount_won BIGINT,
+    total_amount_bet NUMERIC,
+    total_amount_won NUMERIC,
     avg_bet_per_line NUMERIC(20,2),
     win_rate NUMERIC(5,2),
     house_edge NUMERIC(5,2),
